requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
sqlparse==0.5.5
tzdata==2025.3
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-decouple==4.8.0
